clean up

+++ src/main/java/com/example/demo/redis/ExampleProjectRepositoryImpl.java
@@ -2,11 +2,10 @@ package com.example.demo.redis;
 
 import com.redis.om.spring.metamodel.MetamodelField;
 import com.redis.om.spring.search.stream.EntityStream;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Pageable;
-
 import java.util.Set;
 import java.util.stream.Collectors;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 
 public class ExampleProjectRepositoryImpl implements ExampleProjectRepository {
 
@@ -16,7 +15,7 @@ public class ExampleProjectRepositoryImpl implements ExampleProjectRepository {
     @Override
     public <E> Set<E> findAllFields(Class<E> clazz, Pageable pageable, MetamodelField<E, ?>... fields) {
         Set<E> collect = entityStream.of(clazz)
-                .skip(pageable.getPageNumber() * pageable.getPageSize()) // page number
+                .skip((long) pageable.getPageNumber() * pageable.getPageSize()) // page number
                 .limit(pageable.getPageSize()) // page size
                 .project(fields)
                 .collect(Collectors.toSet());
